add: StockVal
- change 'numeric' to 'bigint' in 002_table_stock
- add StockVal file
- add mod.rs file for ds/generic
- use StockVal and Company in global_state.rs
- make application buildable

+++ src/libtrader/db/sql/002_table_stock.sql
@@ -2,7 +2,7 @@ CREATE TABLE asset_schema.stock_vals (
 	id					BIGSERIAL PRIMARY KEY,
 	isin				text NOT NULL,
 	time_since_epoch 	timestamp NOT NULL,
-	ask_price			numeric NOT NULL,
-	bid_price			numeric NOT NULL,
+	ask_price			bigint NOT NULL,
+	bid_price			bigint NOT NULL,
 	volume				bigint NOT NULL
 )
